+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added new command `create test-operator-release` that takes the latest release and creates a new one with the latest version of the provider operator.
+
 ### Changed
 
 - Update to giantswarm/apiextensions 2.6.1.